+++ meridian/constants.py
@@ -433,7 +433,7 @@ FONT_ROBOTO = 'Roboto'
 FONT_GOOGLE_SANS_DISPLAY = 'Google Sans Display'
 
 
-# Default number of max draws per chain in Analyzer.expected_impact()
+# Default number of max draws per chain in Analyzer.expected_outcome()
 DEFAULT_BATCH_SIZE = 100